requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
aiohappyeyeballs==2.4.0
    # via aiohttp
aiohttp==3.10.5
    # via -r requirements.in
aiosignal==1.3.1
    # via aiohttp
attrs==24.2.0
    # via aiohttp
beautifulsoup4==4.12.3
    # via -r requirements.in
certifi==2024.7.4
    # via requests
charset-normalizer==3.3.2
    # via requests
filelock==3.15.4
    # via
    #   huggingface-hub
    #   transformers
frozenlist==1.4.1
    # via
    #   aiohttp
    #   aiosignal
fsspec==2024.6.1
    # via huggingface-hub
huggingface-hub==0.24.6
    # via
    #   tokenizers
    #   transformers
idna==3.7
    # via
    #   requests
    #   yarl
multidict==6.0.5
    # via
    #   aiohttp
    #   yarl
numpy==2.1.0
    # via
    #   pandas
    #   transformers
packaging==24.1
    # via
    #   huggingface-hub
    #   transformers
pandas==2.2.2
    # via -r requirements.in
python-dateutil==2.9.0.post0
    # via pandas
pytz==2024.1
    # via pandas
pyyaml==6.0.2
    # via
    #   -r requirements.in
    #   huggingface-hub
    #   transformers
regex==2024.7.24
    # via transformers
requests==2.32.3
    # via
    #   huggingface-hub
    #   transformers
safetensors==0.4.4
    # via transformers
six==1.16.0
    # via python-dateutil
soupsieve==2.6
    # via beautifulsoup4
tokenizers==0.19.1
    # via transformers
tqdm==4.66.5
    # via
    #   -r requirements.in
    #   huggingface-hub
    #   transformers
transformers==4.44.1
    # via -r requirements.in
typing-extensions==4.12.2
    # via huggingface-hub
tzdata==2024.1
    # via pandas
urllib3==2.2.2
    # via requests
yarl==1.9.4
    # via aiohttp
